Add preprocessor def for runner implementation
Apparently this library has been fundamentally broken as it didn't allow
for including "tests.h" in multiple files.

This approach solves the issue. In short, the "tests.h" is really a
header and source file in one. The implementation is now #ifdef fenced
and will only be included if `TESTS_H_IMPLEMENTATION` is defined before
including `tests.h`. See the `README` for more details.